requirements.txt
# ══════════════════════════════════════════════════════════════════════
# 主人公检测项目依赖
# Protagonist Detection in Movie Clips — Full Requirements
# ══════════════════════════════════════════════════════════════════════

# ── 数据加载（loader.py）──────────────────────────────────────────────
# YT8M TFRecord 解析，使用 SequenceExample 格式
tensorflow-cpu>=2.13.0
# 有 GPU 时改用：tensorflow>=2.13.0

# ── 深度学习框架（multimodal_fusion.py / loss_functions.py）──────────
torch>=2.1.0
# MultiheadAttention / TransformerEncoderLayer / LayerNorm
# AdamW optimizer / CosineAnnealingLR scheduler

# ── 科学计算（全局依赖）─────────────────────────────────────────────
numpy>=1.24.0

# ── 推理聚类（inference.py）─────────────────────────────────────────
scikit-learn>=1.3.0
# KMeans: find_protagonist_cluster
# average_precision_score / cohen_kappa_score: evaluate.py
# precision_score / recall_score / f1_score / confusion_matrix

# ── Excel 读写（generate_annotation_excel.py / evaluate.py / ablation.py）
openpyxl>=3.1.0
# Workbook / PatternFill / Font / Alignment / Border

# pandas 用于 evaluate.py 读取标注表和数据对齐
pandas>=2.0.0

# ── 可视化（visualization/overlay_results.py）───────────────────────
matplotlib>=3.7.0     # 时间轴热图 / 概率曲线
opencv-python>=4.8.0  # 视频帧叠加（overlay_on_video）

# ── 视频下载（visualization/download_mp4.sh 调用）───────────────────
yt-dlp>=2024.1.0

# ── 进度显示（可选）──────────────────────────────────────────────────
tqdm>=4.66.0
